Added casting, to be safe later

+++ src/main.cpp
@@ -10,7 +10,7 @@
 float getTempC(byte pin, int offset = 0, float maxVoltage = 3.3, int resolution = 4096)
 {
     uint16_t read = analogRead(TMP36) + offset;
-    float voltage = read * (maxVoltage / resolution);
+    float voltage = (float)read * (maxVoltage / resolution);
     float temperatureC = (voltage - 0.4) * 100.0; // -40°C à 125°C
     return temperatureC;
 }
@@ -35,7 +35,7 @@ void setup()
 void loop()
 {
     float temp = getTempC(TMP36);
-    
+
     Serial.print(" Temperature (C): ");
     Serial.println(temp);
